systemd: Disable the watchdog feature for NanoPC-T4
The NanoPC-T4 has CONFIG_DW_WATCHDOG as built-in.

When using systemd to pet the watchdog in addition to the above kernel
watchdog driver, it has been observed that after a random period of time
the kernel will complain:

 dw_wdt: keepalive missed, machine will reset

To work around this we disable the systemd watchdog feature and we rely
solely on the in-kernel watchdog keepalive mechanism.

Changelog-entry: Disable the systemd watchdog keep-alive in favor of the in-kernel mechanism (CONFIG_DW_WATCHDOG)
Signed-off-by: Florin Sarbu <florin@balena.io>